ai_services/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
requests
scikit-learn
numpy
python-multipart
Pillow
